Change autosetupmerge to be 'inherit'
This value matches better with how I try to use branches. I almost
always want new branches to have upstreams, but I usually want to use
the parent branch's upstream as my upstream, rather than using the
parent as the upstream (as is the case when using the 'always' option).